NONE: Refactor components
